cdk/requirements.txt
aws-cdk-lib==2.131.0
constructs>=10.0.0,<11.0.0
cdk-ecr-deployment>=3.0.31